+++ crates/bin/disolv-v2x/src/v2x/bucket.rs
@@ -18,8 +18,8 @@ use crate::models::message::{DataBlob, DataType, MessageType, TxMetrics};
 use crate::models::message::PayloadInfo;
 use crate::models::network::{Slice, V2XSlice};
 use crate::models::output::OutputWriter;
-use crate::output::net::NetStatWriter;
-use crate::output::tx::TxDataWriter;
+use crate::out::net::NetStatWriter;
+use crate::out::tx::TxDataWriter;
 use crate::v2x::device::DeviceInfo;
 use crate::v2x::linker::Linker;
 use crate::v2x::space::{Mapper, Space};
